♻️ Rename the module to be more clear its place.
