+++ gearmeshing_ai/mcp_client/gateway_api/__init__.py
@@ -1,6 +1,6 @@
 from .client import GatewayApiClient
 from .errors import GatewayApiError, GatewayServerNotFoundError
-from .models import GatewayServer, GatewayServerCreate, GatewayTransport
+from .domain import GatewayServer, GatewayServerCreate, GatewayTransport
 
 __all__ = [
     "GatewayApiClient",
