+++ test/unit_test/mcp_client/gateway_api/test_gateway_api_client.py
@@ -7,7 +7,7 @@ import pytest
 
 from gearmeshing_ai.mcp_client.gateway_api.client import GatewayApiClient
 from gearmeshing_ai.mcp_client.gateway_api.errors import GatewayServerNotFoundError
-from gearmeshing_ai.mcp_client.gateway_api.models import (
+from gearmeshing_ai.mcp_client.gateway_api.domain import (
     GatewayServerCreate,
     GatewayTransport,
 )
